Move initial test files to providers
