+++ client/src/components/providers/History/HistoryContentProvider/HistoryContentProvider.test.js
@@ -1,6 +1,6 @@
 import { shallowMount } from "@vue/test-utils";
 import { watchUntil, getLocalVue } from "jest/helpers";
-import { sameVueObj, payloadChange } from "components/History/test/providerTestHelpers";
+import { sameVueObj, payloadChange } from "components/providers/History/test/providerTestHelpers";
 import HistoryContentProvider from "./HistoryContentProvider";
 import { SearchParams } from "components/providers/History/SearchParams";
 import { bulkCacheContent, wipeDatabase } from "components/providers/History/caching";
