+++ client/src/components/providers/History/caching/db/changes.test.js
@@ -15,7 +15,7 @@ import { wait } from "jest/helpers";
 
 // test data
 import historyContent from "components/History/test/json/historyContent.json";
-import collectionContent from "components/History/test/json/collectionContent.json";
+import collectionContent from "components/providers/History/test/json/collectionContent.json";
 
 // https://github.com/hirezio/observer-spy/blob/master/README.md
 import { ObserverSpy } from "@hirez_io/observer-spy";
