+++ client/src/components/providers/History/caching/db/find.test.js
@@ -9,7 +9,7 @@ import { find } from "./find";
 
 // test data
 import historyContent from "components/History/test/json/historyContent.json";
-import collectionContent from "components/History/test/json/collectionContent.json";
+import collectionContent from "components/providers/History/test/json/collectionContent.json";
 
 beforeEach(wipeDatabase);
 afterEach(wipeDatabase);
