+++ client/src/components/providers/History/caching/db/monitorQuery.test.js
@@ -10,7 +10,7 @@ import { cacheContent, cacheCollectionContent, bulkCacheContent, bulkCacheDscCon
 
 // test data
 import historyContent from "components/History/test/json/historyContent.json";
-import collectionContent from "components/History/test/json/collectionContent.json";
+import collectionContent from "components/providers/History/test/json/collectionContent.json";
 
 jest.mock("app");
 jest.mock("../../caching");
